Merge iOS 15.x release candidates into the main json
Delete *-RC.json files from iOS 15.x and put the RC information in the main
json file instead, under a createDuplicateEntries key.

Also, some information (like some file sizes/hashes) was present in the -RC
but not in the main json, so here I'm adding that to the main json as well.