sort_os_files: if processing a file fails, say which file
If processing a file throws an exception, for example due to a JSON file
having a syntax error, sort_os_files.py was showing the exception backtrace
but there was no indication of which JSON file caused it. Now we print the
file path, before re-raising the exception so it shows the backtrace as
usual.