TEMP: List Zebra for TNS
This is a weird case where a proper implementation isn't reasonable right now.

TNS on the guide is switching to point to tns-sockport and MeridianFix (TNS). It doesn't make sense to list these as completely separate (they're both just forks of TotallyNotSpyware to some degree).

This wouldn't normally be the biggest deal, however both of these implementations use Zebra. Therefore for the sake of the guide (since this is a high priority matter), let's change these instances to Zebra for now, and a better solution for AppleDB can be done at a later time.